backend/requirements.txt

fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pandas==2.2.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.8
aiofiles==23.1.0
pytest==7.4.0
httpx==0.24.1
email-validator
